move atomic-rating to common
https://coveord.atlassian.net/browse/KIT-3053
