+++ packages/atomic/src/components/search/atomic-breadbox/atomic-breadbox.pcss
@@ -1,5 +1,5 @@
 @import '../../../global/global.pcss';
-@import '../atomic-rating/atomic-rating.pcss';
+@import '../../common/atomic-rating/atomic-rating.pcss';
 
 .max-w-snippet {
   max-width: 30ch;
