--- packages/atomic/src/components/search/atomic-rating/atomic-rating.pcss
@@ -1,23 +1,0 @@
-/**
- * @prop --atomic-rating-icon-active-color: Color of the icon when active.
- */
-.icon-active {
-  color: var(--atomic-rating-icon-active-color, #f6ce3c);
-}
-
-/**
- * @prop --atomic-rating-icon-inactive-color: Color of the icon when inactive.
- */
-.icon-inactive {
-  color: var(--atomic-rating-icon-inactive-color, var(--atomic-neutral));
-}
-
-/**
- * @prop --atomic-rating-icon-outline: Color of the icon's outline.
- */
-.icon-active,
-.icon-inactive {
-  & svg path {
-    stroke: var(--atomic-rating-icon-outline, none);
-  }
-}
