+++ packages/atomic/src/components/search/facets/atomic-rating-facet/atomic-rating-facet.pcss
@@ -1,3 +1,3 @@
 @import '../../../common/facets/facet-common.pcss';
 @import '../../../common/facets/facet-value-checkbox/facet-value-checkbox.pcss';
-@import '../../atomic-rating/atomic-rating.pcss';
+@import '../../../common/atomic-rating/atomic-rating.pcss';
