+++ packages/atomic/src/components/search/facets/atomic-rating-range-facet/atomic-rating-range-facet.pcss
@@ -1,2 +1,2 @@
 @import '../../../common/facets/facet-common.pcss';
-@import '../../atomic-rating/atomic-rating.pcss';
+@import '../../../common/atomic-rating/atomic-rating.pcss';
