+++ packages/atomic/src/components/search/result-template-components/atomic-result-rating/atomic-result-rating.pcss
@@ -1,2 +1,2 @@
 @import '../../../../global/global.pcss';
-@import '../../atomic-rating/atomic-rating.pcss';
+@import '../../../common/atomic-rating/atomic-rating.pcss';
